+++ NEWS.md
@@ -1,5 +1,7 @@
 # DES Patient Pathway 0.1.0.9000
 
+* Added Outpatient "did not attend" (DNA) rate to the model
+
 ---
 
 # DES Patient Pathway 0.1.0