+++ NEWS.md
@@ -1,7 +1,8 @@
-# DES Patient Pathway 0.1.0.9000
+# DES Patient Pathway (development version)
 
 * Added Outpatient "did not attend" (DNA) rate to the model.
 * Removed the probablity distribution (variation) from OP clinic appointment times.  OP clinics are now assumed to be a regular consistent length (adjustable by the user).
+* Added a tab to show this changelog within the app.
 
 ---